further work on decimal

+++ Libraries/01-Shared/Elysium.Core.Template/Decimal.hpp
@@ -100,7 +100,7 @@ namespace Elysium::Core::Template::Math
 
 	inline constexpr Decimal::Decimal(const double Value) noexcept
 		: _HighPart(Elysium::Core::Template::Numeric::NumericTraits<double>::GetValuePreDecimalPoint(Value)),
-		_LowPart(Elysium::Core::Template::Numeric::NumericTraits<float>::GetValuePostDecimalPoint(Value, 3))
+		_LowPart(Elysium::Core::Template::Numeric::NumericTraits<double>::GetValuePostDecimalPoint(Value, 3))
 	{ }
 
 	inline constexpr Decimal::Decimal(const Elysium::Core::Template::System::int64_t HighPart, const Elysium::Core::Template::System::uint64_t LowPart) noexcept
